+++ requirements.txt
@@ -4,4 +4,8 @@ djangorestframework==3.12.0
 sqlparse==0.4.4
 tzdata==2023.3
 nasdaq-data-link==1.0.4
-pandas
+pandas==2.0.1
+scipy==1.10.1
+seaborn==0.12.2
+pymysql==1.0.3
+mayplotlib==3.7.1
