+++ test.py
@@ -6,6 +6,15 @@ from datetime import datetime, timedelta, date
 from PriceDashBoard.models import MaterialsModel, MaterialsPriceModel
 import pandas as pd
 import unittest
+from PriceDashBoard.data_preprocess import data_preprocess
+
+
+class data_preprocess_test(unittest.TestCase):
+    def test_add(self):
+        expected = {"material_name_id":1.0,"price":1995.4,"upper_20":0.0,"upper_60":1.0,"regular_array":1.0}
+        output = data_preprocess(1)
+
+        self.assertEqual(output, expected)
 
 class GetAPITest(unittest.TestCase):
     def test_get_price_data(self):
